+++ cmp/grid/GridModel.js
@@ -142,8 +142,18 @@ export class GridModel extends HoistModel {
     @observable emptyText;
     /** @member {TreeStyle} */
     @observable treeStyle;
-    /** @member {boolean} */
+
+    /** @member {boolean} - Flag to track inline editing at a granular level.
+     *      Will toggle each time row or cell editing is activated or ended.
+     */
     @observable isEditing = false;
+
+    /** @member {boolean} - Flag to track inline editing at a general level.
+     *      Will not change during transient navigation from cell to cell or row to row,
+     *      but rather is debounced such that grid editing will need to "settle" for a
+     *      short time before toggling.
+     */
+    @observable isInEditingMode = false;
 
     static defaultContextMenu = [
         'filter',
@@ -425,6 +435,12 @@ export class GridModel extends HoistModel {
         this.onCellClicked = onCellClicked;
         this.onCellDoubleClicked = onCellDoubleClicked;
         this.onCellContextMenu = onCellContextMenu;
+
+        this.addReaction({
+            track: () => this.isEditing,
+            run: (isEditing) => this.isInEditingMode = isEditing,
+            debounce: 500
+        });
     }
 
     /**